refactor: 체스보드 초기화 역할 분리
